面倒なのでコピーファイルを1つにする

+++ copy.sh
@@ -1,10 +1,16 @@
 #!/bin/sh
 
 # 実行ディレクトリが変わってしまうため、シンボリックリンクではなくコピーする
-rm -r ./gulpfile.js
-cp -r ../gulpfile.js ./gulpfile.js
+if [ -e ./gulpfile.js ]; then
+  rm -r ./gulpfile.js
+  echo "gulpfile.jsを削除しました"
+fi
+cp ../gulpfile.js ./gulpfile.js
 echo "gulpfile.jsのコピーが完了しました"
 
-rm -r ./lib/
+if [ -e ./lib/ ]; then
+  rm -r ./lib/
+  echo "libを削除しました"
+fi
 cp -r ../lib/ ./lib/
 echo "libのコピーが完了しました"
